requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
python-dotenv
chromadb
llama-index
llama-index-vector-stores-chroma
aiohttp